pyproject.toml
[tool.poetry]
name = "rfidtools"
version = "2.0.0"
description = "A tool for RFID production at CIOT."
authors = ["ar-bebop <aribeir4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
pyodbc = "^4.0.39"
requests = "^2.31.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
